refactor: Remove code from `components/stacks-codec`
