+++ components/stacks-codec/src/codec.rs
@@ -1,4 +1,4 @@
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryFrom;
 use std::fmt;
 use std::io::{Read, Write};
 use std::ops::{Deref, DerefMut};
@@ -9,8 +9,7 @@ use clarity::codec::{
     read_next, read_next_exact, write_next, Error as CodecError, MAX_MESSAGE_LEN,
 };
 use clarity::types::chainstate::{
-    BlockHeaderHash, BurnchainHeaderHash, ConsensusHash, StacksAddress, StacksBlockId,
-    StacksPublicKey, StacksWorkScore, TrieHash,
+    BlockHeaderHash, ConsensusHash, StacksAddress, StacksBlockId, StacksPublicKey, TrieHash,
 };
 use clarity::types::{PrivateKey, StacksEpochId};
 use clarity::vm::types::{
@@ -2127,186 +2126,6 @@ impl StacksTransaction {
         &self.auth
     }
 
-    /// begin signing the transaction.
-    /// If this is a sponsored transaction, then the origin only commits to knowing that it is
-    /// sponsored.  It does _not_ commit to the sponsored fields, so set them all to sentinel
-    /// values.
-    /// Return the initial sighash.
-    fn sign_begin(&self) -> Txid {
-        let mut tx = self.clone();
-        tx.auth = tx.auth.into_initial_sighash_auth();
-        tx.txid()
-    }
-
-    /// begin verifying a transaction.
-    /// return the initial sighash
-    fn verify_begin(&self) -> Txid {
-        let mut tx = self.clone();
-        tx.auth = tx.auth.into_initial_sighash_auth();
-        tx.txid()
-    }
-
-    /// Sign a sighash and append the signature and public key to the given spending condition.
-    /// Returns the next sighash
-    fn sign_and_append(
-        condition: &mut TransactionSpendingCondition,
-        cur_sighash: &Txid,
-        auth_flag: &TransactionAuthFlags,
-        privk: &Secp256k1PrivateKey,
-    ) -> Result<Txid, CodecError> {
-        let (next_sig, next_sighash) = TransactionSpendingCondition::next_signature(
-            cur_sighash,
-            auth_flag,
-            condition.tx_fee(),
-            condition.nonce(),
-            privk,
-        )?;
-        match condition {
-            TransactionSpendingCondition::Singlesig(ref mut cond) => {
-                cond.set_signature(next_sig);
-                Ok(next_sighash)
-            }
-            TransactionSpendingCondition::Multisig(ref mut cond) => {
-                cond.push_signature(
-                    if privk.compress_public() {
-                        TransactionPublicKeyEncoding::Compressed
-                    } else {
-                        TransactionPublicKeyEncoding::Uncompressed
-                    },
-                    next_sig,
-                );
-                Ok(next_sighash)
-            }
-            TransactionSpendingCondition::OrderIndependentMultisig(ref mut cond) => {
-                cond.push_signature(
-                    if privk.compress_public() {
-                        TransactionPublicKeyEncoding::Compressed
-                    } else {
-                        TransactionPublicKeyEncoding::Uncompressed
-                    },
-                    next_sig,
-                );
-                Ok(*cur_sighash)
-            }
-        }
-    }
-
-    /// Pop the last auth field
-    fn pop_auth_field(
-        condition: &mut TransactionSpendingCondition,
-    ) -> Option<TransactionAuthField> {
-        match condition {
-            TransactionSpendingCondition::Multisig(ref mut cond) => cond.pop_auth_field(),
-            TransactionSpendingCondition::OrderIndependentMultisig(ref mut cond) => {
-                cond.pop_auth_field()
-            }
-            TransactionSpendingCondition::Singlesig(ref mut cond) => cond.pop_signature(),
-        }
-    }
-
-    /// Append a public key to a multisig condition
-    fn append_pubkey(
-        condition: &mut TransactionSpendingCondition,
-        pubkey: &StacksPublicKey,
-    ) -> Result<(), CodecError> {
-        match condition {
-            TransactionSpendingCondition::Multisig(ref mut cond) => {
-                cond.push_public_key(pubkey.clone());
-                Ok(())
-            }
-            TransactionSpendingCondition::OrderIndependentMultisig(ref mut cond) => {
-                cond.push_public_key(pubkey.clone());
-                Ok(())
-            }
-            _ => Err(CodecError::SigningError(
-                "Not a multisig condition".to_string(),
-            )),
-        }
-    }
-
-    /// Append the next signature from the origin account authorization.
-    /// Return the next sighash.
-    pub fn sign_next_origin(
-        &mut self,
-        cur_sighash: &Txid,
-        privk: &Secp256k1PrivateKey,
-    ) -> Result<Txid, CodecError> {
-        let next_sighash = match self.auth {
-            TransactionAuth::Standard(ref mut origin_condition)
-            | TransactionAuth::Sponsored(ref mut origin_condition, _) => {
-                StacksTransaction::sign_and_append(
-                    origin_condition,
-                    cur_sighash,
-                    &TransactionAuthFlags::AuthStandard,
-                    privk,
-                )?
-            }
-        };
-        Ok(next_sighash)
-    }
-
-    /// Append the next public key to the origin account authorization.
-    pub fn append_next_origin(&mut self, pubk: &StacksPublicKey) -> Result<(), CodecError> {
-        match self.auth {
-            TransactionAuth::Standard(ref mut origin_condition) => {
-                StacksTransaction::append_pubkey(origin_condition, pubk)
-            }
-            TransactionAuth::Sponsored(ref mut origin_condition, _) => {
-                StacksTransaction::append_pubkey(origin_condition, pubk)
-            }
-        }
-    }
-
-    /// Append the next signature from the sponsoring account.
-    /// Return the next sighash
-    pub fn sign_next_sponsor(
-        &mut self,
-        cur_sighash: &Txid,
-        privk: &Secp256k1PrivateKey,
-    ) -> Result<Txid, CodecError> {
-        let next_sighash = match self.auth {
-            TransactionAuth::Standard(_) => {
-                // invalid
-                return Err(CodecError::SigningError(
-                    "Cannot sign standard authorization with a sponsoring private key".to_string(),
-                ));
-            }
-            TransactionAuth::Sponsored(_, ref mut sponsor_condition) => {
-                StacksTransaction::sign_and_append(
-                    sponsor_condition,
-                    cur_sighash,
-                    &TransactionAuthFlags::AuthSponsored,
-                    privk,
-                )?
-            }
-        };
-        Ok(next_sighash)
-    }
-
-    /// Append the next public key to the sponsor account authorization.
-    pub fn append_next_sponsor(&mut self, pubk: &StacksPublicKey) -> Result<(), CodecError> {
-        match self.auth {
-            TransactionAuth::Standard(_) => Err(CodecError::SigningError(
-                "Cannot append a public key to the sponsor of a standard auth condition"
-                    .to_string(),
-            )),
-            TransactionAuth::Sponsored(_, ref mut sponsor_condition) => {
-                StacksTransaction::append_pubkey(sponsor_condition, pubk)
-            }
-        }
-    }
-
-    /// Verify this transaction's signatures
-    pub fn verify(&self) -> Result<(), CodecError> {
-        self.auth.verify(&self.verify_begin())
-    }
-
-    /// Verify the transaction's origin signatures only.
-    /// Used by sponsors to get the next sig-hash to sign.
-    pub fn verify_origin(&self) -> Result<Txid, CodecError> {
-        self.auth.verify_origin(&self.verify_begin())
-    }
-
     /// Get the origin account's address
     pub fn origin_address(&self) -> StacksAddress {
         match (&self.version, &self.auth) {
@@ -2494,203 +2313,6 @@ impl Txid {
     }
 }
 
-#[derive(Debug, Clone, PartialEq)]
-pub struct StacksTransactionSigner {
-    pub tx: StacksTransaction,
-    pub sighash: Txid,
-    origin_done: bool,
-    check_oversign: bool,
-    check_overlap: bool,
-}
-
-impl StacksTransactionSigner {
-    pub fn new(tx: &StacksTransaction) -> StacksTransactionSigner {
-        StacksTransactionSigner {
-            tx: tx.clone(),
-            sighash: tx.sign_begin(),
-            origin_done: false,
-            check_oversign: true,
-            check_overlap: true,
-        }
-    }
-
-    pub fn new_sponsor(
-        tx: &StacksTransaction,
-        spending_condition: TransactionSpendingCondition,
-    ) -> Result<StacksTransactionSigner, CodecError> {
-        if !tx.auth.is_sponsored() {
-            return Err(CodecError::GenericError(
-                "IncompatibleSpendingConditionError".into(),
-            ));
-        }
-        let mut new_tx = tx.clone();
-        new_tx.auth.set_sponsor(spending_condition)?;
-        let origin_sighash = new_tx.verify_origin()?;
-
-        Ok(StacksTransactionSigner {
-            tx: new_tx,
-            sighash: origin_sighash,
-            origin_done: true,
-            check_oversign: true,
-            check_overlap: true,
-        })
-    }
-
-    pub fn resume(&mut self, tx: &StacksTransaction) {
-        self.tx = tx.clone()
-    }
-
-    pub fn disable_checks(&mut self) {
-        self.check_oversign = false;
-        self.check_overlap = false;
-    }
-
-    pub fn sign_origin(&mut self, privk: &Secp256k1PrivateKey) -> Result<(), CodecError> {
-        if self.check_overlap && self.origin_done {
-            // can't sign another origin private key since we started signing sponsors
-            return Err(CodecError::SigningError(
-                "Cannot sign origin after sponsor key".to_string(),
-            ));
-        }
-
-        match self.tx.auth {
-            TransactionAuth::Standard(ref origin_condition) => {
-                if self.check_oversign
-                    && origin_condition.num_signatures() >= origin_condition.signatures_required()
-                {
-                    return Err(CodecError::SigningError(
-                        "Origin would have too many signatures".to_string(),
-                    ));
-                }
-            }
-            TransactionAuth::Sponsored(ref origin_condition, _) => {
-                if self.check_oversign
-                    && origin_condition.num_signatures() >= origin_condition.signatures_required()
-                {
-                    return Err(CodecError::SigningError(
-                        "Origin would have too many signatures".to_string(),
-                    ));
-                }
-            }
-        }
-
-        let next_sighash = self.tx.sign_next_origin(&self.sighash, privk)?;
-        self.sighash = next_sighash;
-        Ok(())
-    }
-
-    pub fn append_origin(&mut self, pubk: &Secp256k1PublicKey) -> Result<(), CodecError> {
-        if self.check_overlap && self.origin_done {
-            // can't append another origin key
-            return Err(CodecError::SigningError(
-                "Cannot append public key to origin after sponsor key".to_string(),
-            ));
-        }
-
-        self.tx.append_next_origin(pubk)
-    }
-
-    pub fn sign_sponsor(&mut self, privk: &Secp256k1PrivateKey) -> Result<(), CodecError> {
-        match self.tx.auth {
-            TransactionAuth::Sponsored(_, ref sponsor_condition) => {
-                if self.check_oversign
-                    && sponsor_condition.num_signatures() >= sponsor_condition.signatures_required()
-                {
-                    return Err(CodecError::SigningError(
-                        "Sponsor would have too many signatures".to_string(),
-                    ));
-                }
-            }
-            TransactionAuth::Standard(_) => todo!(),
-        }
-
-        let next_sighash = self.tx.sign_next_sponsor(&self.sighash, privk)?;
-        self.sighash = next_sighash;
-        self.origin_done = true;
-        Ok(())
-    }
-
-    pub fn append_sponsor(&mut self, pubk: &Secp256k1PublicKey) -> Result<(), CodecError> {
-        self.tx.append_next_sponsor(pubk)
-    }
-
-    pub fn pop_origin_auth_field(&mut self) -> Option<TransactionAuthField> {
-        match self.tx.auth {
-            TransactionAuth::Standard(ref mut origin_condition) => {
-                StacksTransaction::pop_auth_field(origin_condition)
-            }
-            TransactionAuth::Sponsored(ref mut origin_condition, _) => {
-                StacksTransaction::pop_auth_field(origin_condition)
-            }
-        }
-    }
-
-    pub fn pop_sponsor_auth_field(&mut self) -> Option<TransactionAuthField> {
-        match self.tx.auth {
-            TransactionAuth::Sponsored(_, ref mut sponsor_condition) => {
-                StacksTransaction::pop_auth_field(sponsor_condition)
-            }
-            _ => None,
-        }
-    }
-
-    pub fn complete(&self) -> bool {
-        match self.tx.auth {
-            TransactionAuth::Standard(ref origin_condition) => {
-                origin_condition.num_signatures() >= origin_condition.signatures_required()
-            }
-            TransactionAuth::Sponsored(ref origin_condition, ref sponsored_condition) => {
-                origin_condition.num_signatures() >= origin_condition.signatures_required()
-                    && sponsored_condition.num_signatures()
-                        >= sponsored_condition.signatures_required()
-                    && (self.origin_done || !self.check_overlap)
-            }
-        }
-    }
-
-    pub fn get_tx_incomplete(&self) -> StacksTransaction {
-        self.tx.clone()
-    }
-
-    pub fn get_tx(&self) -> Option<StacksTransaction> {
-        if self.complete() {
-            Some(self.tx.clone())
-        } else {
-            None
-        }
-    }
-}
-
-/// A block that contains blockchain-anchored data
-/// (corresponding to a LeaderBlockCommitOp)
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct StacksBlock {
-    pub header: StacksBlockHeader,
-    pub txs: Vec<StacksTransaction>,
-}
-
-/// A microblock that contains non-blockchain-anchored data,
-/// but is tied to an on-chain block
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct StacksMicroblock {
-    pub header: StacksMicroblockHeader,
-    pub txs: Vec<StacksTransaction>,
-}
-
-/// The header for an on-chain-anchored Stacks block
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct StacksBlockHeader {
-    pub version: u8,
-    pub total_work: StacksWorkScore, // NOTE: this is the work done on the chain tip this block builds on (i.e. take this from the parent)
-    pub proof: VRFProof,
-    pub parent_block: BlockHeaderHash, // NOTE: even though this is also present in the burn chain, we need this here for super-light clients that don't even have burn chain headers
-    pub parent_microblock: BlockHeaderHash,
-    pub parent_microblock_sequence: u16,
-    pub tx_merkle_root: Sha512Trunc256Sum,
-    pub state_index_root: TrieHash,
-    pub microblock_pubkey_hash: Hash160, // we'll get the public key back from the first signature (note that this is the Hash160 of the _compressed_ public key)
-}
-
 /// Header structure for a microblock
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct StacksMicroblockHeader {
@@ -2868,178 +2490,6 @@ impl StacksMessageCodec for NakamotoBlock {
 
         Ok(NakamotoBlock { header, txs })
     }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-/// A vote across the signer set for a block
-pub struct NakamotoBlockVote {
-    pub signer_signature_hash: Sha512Trunc256Sum,
-    pub rejected: bool,
-}
-
-impl StacksMessageCodec for NakamotoBlockVote {
-    fn consensus_serialize<W: std::io::Write>(&self, fd: &mut W) -> Result<(), CodecError> {
-        write_next(fd, &self.signer_signature_hash)?;
-        if self.rejected {
-            write_next(fd, &1u8)?;
-        }
-        Ok(())
-    }
-
-    fn consensus_deserialize<R: std::io::Read>(fd: &mut R) -> Result<Self, CodecError> {
-        let signer_signature_hash = read_next(fd)?;
-        let rejected_byte: Option<u8> = read_next(fd).ok();
-        let rejected = rejected_byte.is_some();
-        Ok(Self {
-            signer_signature_hash,
-            rejected,
-        })
-    }
-}
-
-// values a miner uses to produce the next block
-pub const MINER_BLOCK_CONSENSUS_HASH: ConsensusHash = ConsensusHash([1u8; 20]);
-pub const MINER_BLOCK_HEADER_HASH: BlockHeaderHash = BlockHeaderHash([1u8; 32]);
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum StacksBlockHeaderTypes {
-    Epoch2(StacksBlockHeader),
-    Nakamoto(NakamotoBlockHeader),
-}
-
-impl From<StacksBlockHeader> for StacksBlockHeaderTypes {
-    fn from(value: StacksBlockHeader) -> Self {
-        Self::Epoch2(value)
-    }
-}
-
-impl From<NakamotoBlockHeader> for StacksBlockHeaderTypes {
-    fn from(value: NakamotoBlockHeader) -> Self {
-        Self::Nakamoto(value)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct StacksHeaderInfo {
-    /// Stacks block header
-    pub anchored_header: StacksBlockHeaderTypes,
-    /// Last microblock header (Stacks 2.x only; this is None in Stacks 3.x)
-    pub microblock_tail: Option<StacksMicroblockHeader>,
-    /// Height of this Stacks block
-    pub stacks_block_height: u64,
-    /// MARF root hash of the headers DB (not consensus critical)
-    pub index_root: TrieHash,
-    /// consensus hash of the burnchain block in which this miner was selected to produce this block
-    pub consensus_hash: ConsensusHash,
-    /// Hash of the burnchain block in which this miner was selected to produce this block
-    pub burn_header_hash: BurnchainHeaderHash,
-    /// Height of the burnchain block
-    pub burn_header_height: u32,
-    /// Timestamp of the burnchain block
-    pub burn_header_timestamp: u64,
-    /// Size of the block corresponding to `anchored_header` in bytes
-    pub anchored_block_size: u64,
-    /// The burnchain tip that is passed to Clarity while processing this block.
-    /// This should always be `Some()` for Nakamoto blocks and `None` for 2.x blocks
-    pub burn_view: Option<ConsensusHash>,
-}
-
-/// A record of a coin reward for a miner.  There will be at most two of these for a miner: one for
-/// the coinbase + block-txs + confirmed-mblock-txs, and one for the produced-mblock-txs.  The
-/// latter reward only stores the produced-mblock-txs, and is only ever stored if the microblocks
-/// are ever confirmed.
-#[derive(Debug, Clone, PartialEq)]
-pub struct MinerReward {
-    /// address of the miner that produced the block
-    pub address: StacksAddress,
-    /// address of the entity that receives the block reward.
-    /// Ignored pre-2.1
-    pub recipient: PrincipalData,
-    /// block coinbase
-    pub coinbase: u128,
-    /// block transaction fees
-    pub tx_fees_anchored: u128,
-    /// microblock transaction fees from transactions *mined* by this miner
-    pub tx_fees_streamed_produced: u128,
-    /// microblock transaction fees from transactions *confirmed* by this miner
-    pub tx_fees_streamed_confirmed: u128,
-    /// virtual transaction index in the block where these rewards get applied.  the miner's reward
-    /// is applied first (so vtxindex == 0) and user-burn supports would be applied after (so
-    /// vtxindex > 0).
-    pub vtxindex: u32,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct MinerRewardInfo {
-    pub from_block_consensus_hash: ConsensusHash,
-    pub from_stacks_block_hash: BlockHeaderHash,
-    pub from_parent_block_consensus_hash: ConsensusHash,
-    pub from_parent_stacks_block_hash: BlockHeaderHash,
-}
-
-// maximum amount of data a leader can send during its epoch (2MB)
-pub const MAX_EPOCH_SIZE: u32 = 2 * 1024 * 1024;
-
-// maximum microblock size is 64KB, but note that the current leader has a space budget of
-// $MAX_EPOCH_SIZE bytes (so the average microblock size needs to be 4kb if there are 256 of them)
-pub const MAX_MICROBLOCK_SIZE: u32 = 65536;
-
-pub fn build_contract_call_transaction(
-    contract_id: String,
-    function_name: String,
-    args: Vec<Value>,
-    nonce: u64,
-    fee: u64,
-    sender_secret_key: &[u8],
-) -> StacksTransaction {
-    let contract_id =
-        QualifiedContractIdentifier::parse(&contract_id).expect("Contract identifier invalid");
-
-    let payload = TransactionContractCall {
-        address: contract_id.issuer.into(),
-        contract_name: contract_id.name,
-        function_name: function_name.try_into().unwrap(),
-        function_args: args,
-    };
-
-    let secret_key = Secp256k1PrivateKey::from_slice(sender_secret_key).unwrap();
-    let mut public_key = Secp256k1PublicKey::from_private(&secret_key);
-    public_key.set_compressed(true);
-
-    let anchor_mode = TransactionAnchorMode::Any;
-    let signer_addr =
-        StacksAddress::from_public_keys(0, &AddressHashMode::SerializeP2PKH, 1, &vec![public_key])
-            .unwrap();
-
-    let spending_condition = TransactionSpendingCondition::Singlesig(SinglesigSpendingCondition {
-        signer: signer_addr.bytes().clone(),
-        nonce,
-        tx_fee: fee,
-        hash_mode: SinglesigHashMode::P2PKH,
-        key_encoding: TransactionPublicKeyEncoding::Compressed,
-        signature: MessageSignature::empty(),
-    });
-
-    let auth = TransactionAuth::Standard(spending_condition);
-    let unsigned_tx = StacksTransaction {
-        version: TransactionVersion::Testnet,
-        chain_id: 0x80000000, // MAINNET=0x00000001
-        auth,
-        anchor_mode,
-        post_condition_mode: TransactionPostConditionMode::Allow,
-        post_conditions: vec![],
-        payload: TransactionPayload::ContractCall(payload),
-    };
-
-    let mut unsigned_tx_bytes = vec![];
-    unsigned_tx
-        .consensus_serialize(&mut unsigned_tx_bytes)
-        .expect("FATAL: invalid transaction");
-
-    let mut tx_signer = StacksTransactionSigner::new(&unsigned_tx);
-    tx_signer.sign_origin(&secret_key).unwrap();
-
-    tx_signer.get_tx().unwrap()
 }
 
 impl StacksMessageCodec for TransactionContractCall {
