+++ components/stacks-rpc-client/src/crypto.rs
@@ -1,5 +1,3 @@
-use std::str::FromStr;
-
 use clarinet_utils::get_bip32_keys_from_mnemonic;
 use clarity::codec::StacksMessageCodec;
 use clarity::types::chainstate::StacksAddress;
@@ -7,10 +5,16 @@ use clarity::util::secp256k1::{MessageSignature, Secp256k1PrivateKey, Secp256k1P
 use clarity::vm::types::{PrincipalData, QualifiedContractIdentifier};
 use clarity::vm::{ClarityName, ClarityVersion, ContractName, Value as ClarityValue};
 use libsecp256k1::PublicKey;
-use stacks_codec::codec::*;
 use stacks_common::address::{
     AddressHashMode, C32_ADDRESS_VERSION_MAINNET_SINGLESIG, C32_ADDRESS_VERSION_TESTNET_SINGLESIG,
 };
+use stackslib::chainstate::stacks::{
+    SinglesigHashMode, SinglesigSpendingCondition, StacksTransaction, StacksTransactionSigner,
+    TokenTransferMemo, TransactionAnchorMode, TransactionAuth, TransactionContractCall,
+    TransactionPayload, TransactionPostConditionMode, TransactionPublicKeyEncoding,
+    TransactionSmartContract, TransactionSpendingCondition, TransactionVersion,
+};
+use stackslib::util_lib::strings::StacksString;
 
 #[derive(Clone, Debug)]
 pub struct Wallet {
@@ -161,4 +165,62 @@ pub fn encode_contract_publish(
         tx_fee,
         anchor_mode,
     )
+}
+
+pub fn build_contract_call_transaction(
+    contract_id: String,
+    function_name: String,
+    args: Vec<ClarityValue>,
+    nonce: u64,
+    fee: u64,
+    sender_secret_key: &[u8],
+) -> StacksTransaction {
+    let contract_id =
+        QualifiedContractIdentifier::parse(&contract_id).expect("Contract identifier invalid");
+
+    let payload = TransactionContractCall {
+        address: contract_id.issuer.into(),
+        contract_name: contract_id.name,
+        function_name: function_name.try_into().unwrap(),
+        function_args: args,
+    };
+
+    let secret_key = Secp256k1PrivateKey::from_slice(sender_secret_key).unwrap();
+    let mut public_key = Secp256k1PublicKey::from_private(&secret_key);
+    public_key.set_compressed(true);
+
+    let anchor_mode = TransactionAnchorMode::Any;
+    let signer_addr =
+        StacksAddress::from_public_keys(0, &AddressHashMode::SerializeP2PKH, 1, &vec![public_key])
+            .unwrap();
+
+    let spending_condition = TransactionSpendingCondition::Singlesig(SinglesigSpendingCondition {
+        signer: signer_addr.bytes().clone(),
+        nonce,
+        tx_fee: fee,
+        hash_mode: SinglesigHashMode::P2PKH,
+        key_encoding: TransactionPublicKeyEncoding::Compressed,
+        signature: MessageSignature::empty(),
+    });
+
+    let auth = TransactionAuth::Standard(spending_condition);
+    let unsigned_tx = StacksTransaction {
+        version: TransactionVersion::Testnet,
+        chain_id: 0x80000000, // MAINNET=0x00000001
+        auth,
+        anchor_mode,
+        post_condition_mode: TransactionPostConditionMode::Allow,
+        post_conditions: vec![],
+        payload: TransactionPayload::ContractCall(payload),
+    };
+
+    let mut unsigned_tx_bytes = vec![];
+    unsigned_tx
+        .consensus_serialize(&mut unsigned_tx_bytes)
+        .expect("FATAL: invalid transaction");
+
+    let mut tx_signer = StacksTransactionSigner::new(&unsigned_tx);
+    tx_signer.sign_origin(&secret_key).unwrap();
+
+    tx_signer.get_tx().unwrap()
 }
